+++ TeamCode/src/main/java/pedroPathing/tests/localization/SensorGoBildaPinpointExample.java
@@ -20,7 +20,7 @@
  *   SOFTWARE.
  */
 
-package pedroPathing.tuners_tests.localization;
+package pedroPathing.tests.localization;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
